Excluindo níveis com controle

+++ admin/cadastros/niveis/delete.php
@@ -11,6 +11,16 @@ else {
 	die();
 }
 
+// não excluir se tiver usuários
+$sql = sql("SELECT COUNT(id) as total FROM usuarios WHERE nivel_id = $id");
+$total_usuarios = $sql['total'];
+
+if ($total_usuarios > 0) {
+	$_SESSION['alert_warning'] = "Nível com usuários cadastrados, impossível excluir";
+	header("Location: " . URL . '/admin/cadastros/niveis');
+	die();
+}
+
 $sql = sql_delete("DELETE FROM niveis WHERE id = '$id'");
 
 $_SESSION['alert_danger'] = "Nível excluído com sucesso";
